apps/rag_lmbd_embeddings/requirements.txt
boto3
python-dotenv
pdfplumber
langchain
psycopg2-binary
langchain-text-splitters
numpy